pyproject.toml
[project]
name = "deepface-demo"
version = "0.1.0"
description = "This app uses [DeepFace](https://github.com/serengil/deepface) to analyze your face and display the results"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepface>=0.0.93",
    "numpy<2",
    "pillow>=11.1.0",
    "streamlit>=1.41.1",
    "tf-keras>=2.18.0",
]
